fixed all tests that werent passing

+++ server/db/models/product.js
@@ -25,7 +25,7 @@ var Product = db.define('product', {
     defaultValue: 'Miscellaneous'
   },
   price: {
-    type: Sequelize.DOUBLE,
+    type: Sequelize.DECIMAL(10, 2),
     allowNull: false,
     validate: {
       min: 0
